Move OpenClaw telemetry-hal importer from a supported source to an example
The openclaw-telemetry-hal importer is a one-off (written to import CRUX1-style
captures) and doesn't warrant being a first-class Scout source. Relocate it to
examples/sources/openclaw/ as a self-contained, copy-into-your-project example:

- Move the importer modules to examples/sources/openclaw/telemetry_hal/ (imports
  unchanged; depends only on public inspect_scout / inspect_ai APIs).
- Move the test suite + fixtures alongside it (relative imports); it is no longer
  part of CI (testpaths = tests/) but runs via pytest on the example path.
- Move the design doc in as README.md and add a runnable populate_db.py.
- De-register the source from inspect_scout.sources and revert the db_importing /
  sources reference docs, so it no longer appears in `scout import --sources`.
