+++ examples/sources/openclaw_telemetry_hal/.gitignore
@@ -1,0 +1,1 @@
+transcripts/
