+++ examples/sources/openclaw_telemetry_hal/tests/fixtures/crux1-sample-telemetry.jsonl
