+++ src/inspect_scout/sources/__init__.py
@@ -1,15 +1,7 @@
 from ._claude_code import claude_code
 from ._langsmith import langsmith
 from ._logfire import logfire
-from ._openclaw import openclaw_telemetry_hal
 from ._phoenix import phoenix
 from ._weave import weave
 
-__all__ = [
-    "claude_code",
-    "openclaw_telemetry_hal",
-    "phoenix",
-    "langsmith",
-    "logfire",
-    "weave",
-]
+__all__ = ["claude_code", "phoenix", "langsmith", "logfire", "weave"]
