--- src/inspect_scout/sources/_openclaw/__init__.py
@@ -1,14 +1,0 @@
-"""OpenClaw import sources.
-
-Hosts the family of OpenClaw importers. Currently only ``_telemetry_hal`` (the
-``openclaw-telemetry-hal`` JSONL telemetry importer); a native-session importer
-is expected to join it as a sibling subpackage, with its own entry point and
-``source_type``.
-"""
-
-from ._telemetry_hal import (
-    OPENCLAW_TELEMETRY_HAL_SOURCE_TYPE,
-    openclaw_telemetry_hal,
-)
-
-__all__ = ["openclaw_telemetry_hal", "OPENCLAW_TELEMETRY_HAL_SOURCE_TYPE"]
